+++ kaggle/api/kaggle_api.py
@@ -14,6 +14,22 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+#!/usr/bin/python
+#
+# Copyright 2023 Kaggle Inc
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 # coding: utf-8
 
 """
@@ -1732,6 +1748,7 @@ class KaggleApi(object):
         :param async_req bool
         :param str owner_slug: Dataset owner (required)
         :param str dataset_slug: Dataset name (required)
+        :param str dataset_version_number: Dataset version number
         :return: Result
                  If the method is called asynchronously,
                  returns the request thread.
@@ -1754,12 +1771,13 @@ class KaggleApi(object):
         :param async_req bool
         :param str owner_slug: Dataset owner (required)
         :param str dataset_slug: Dataset name (required)
+        :param str dataset_version_number: Dataset version number
         :return: Result
                  If the method is called asynchronously,
                  returns the request thread.
         """
 
-        all_params = ['owner_slug', 'dataset_slug']  # noqa: E501
+        all_params = ['owner_slug', 'dataset_slug', 'dataset_version_number']  # noqa: E501
         all_params.append('async_req')
         all_params.append('_return_http_data_only')
         all_params.append('_preload_content')
@@ -1792,6 +1810,8 @@ class KaggleApi(object):
             path_params['datasetSlug'] = params['dataset_slug']  # noqa: E501
 
         query_params = []
+        if 'dataset_version_number' in params:
+            query_params.append(('datasetVersionNumber', params['dataset_version_number']))  # noqa: E501
 
         header_params = {}
 
@@ -2022,107 +2042,6 @@ class KaggleApi(object):
 
         return self.api_client.call_api(
             '/datasets/upload/file/{contentLength}/{lastModifiedDateUtc}', 'POST',
-            path_params,
-            query_params,
-            header_params,
-            body=body_params,
-            post_params=form_params,
-            files=local_var_files,
-            response_type='Result',  # noqa: E501
-            auth_settings=auth_settings,
-            async_req=params.get('async_req'),
-            _return_http_data_only=params.get('_return_http_data_only'),
-            _preload_content=params.get('_preload_content', True),
-            _request_timeout=params.get('_request_timeout'),
-            collection_formats=collection_formats)
-
-    def datasets_view(self, owner_slug, dataset_slug, **kwargs):  # noqa: E501
-        """Show details about a dataset  # noqa: E501
-
-        This method makes a synchronous HTTP request by default. To make an
-        asynchronous HTTP request, please pass async_req=True
-        >>> thread = api.datasets_view(owner_slug, dataset_slug, async_req=True)
-        >>> result = thread.get()
-
-        :param async_req bool
-        :param str owner_slug: Dataset owner (required)
-        :param str dataset_slug: Dataset name (required)
-        :return: Result
-                 If the method is called asynchronously,
-                 returns the request thread.
-        """
-        kwargs['_return_http_data_only'] = True
-        if kwargs.get('async_req'):
-            return self.datasets_view_with_http_info(owner_slug, dataset_slug, **kwargs)  # noqa: E501
-        else:
-            (data) = self.datasets_view_with_http_info(owner_slug, dataset_slug, **kwargs)  # noqa: E501
-            return data
-
-    def datasets_view_with_http_info(self, owner_slug, dataset_slug, **kwargs):  # noqa: E501
-        """Show details about a dataset  # noqa: E501
-
-        This method makes a synchronous HTTP request by default. To make an
-        asynchronous HTTP request, please pass async_req=True
-        >>> thread = api.datasets_view_with_http_info(owner_slug, dataset_slug, async_req=True)
-        >>> result = thread.get()
-
-        :param async_req bool
-        :param str owner_slug: Dataset owner (required)
-        :param str dataset_slug: Dataset name (required)
-        :return: Result
-                 If the method is called asynchronously,
-                 returns the request thread.
-        """
-
-        all_params = ['owner_slug', 'dataset_slug']  # noqa: E501
-        all_params.append('async_req')
-        all_params.append('_return_http_data_only')
-        all_params.append('_preload_content')
-        all_params.append('_request_timeout')
-
-        params = locals()
-        for key, val in six.iteritems(params['kwargs']):
-            if key not in all_params:
-                raise TypeError(
-                    "Got an unexpected keyword argument '%s'"
-                    " to method datasets_view" % key
-                )
-            params[key] = val
-        del params['kwargs']
-        # verify the required parameter 'owner_slug' is set
-        if ('owner_slug' not in params or
-                params['owner_slug'] is None):
-            raise ValueError("Missing the required parameter `owner_slug` when calling `datasets_view`")  # noqa: E501
-        # verify the required parameter 'dataset_slug' is set
-        if ('dataset_slug' not in params or
-                params['dataset_slug'] is None):
-            raise ValueError("Missing the required parameter `dataset_slug` when calling `datasets_view`")  # noqa: E501
-
-        collection_formats = {}
-
-        path_params = {}
-        if 'owner_slug' in params:
-            path_params['ownerSlug'] = params['owner_slug']  # noqa: E501
-        if 'dataset_slug' in params:
-            path_params['datasetSlug'] = params['dataset_slug']  # noqa: E501
-
-        query_params = []
-
-        header_params = {}
-
-        form_params = []
-        local_var_files = {}
-
-        body_params = None
-        # HTTP header `Accept`
-        header_params['Accept'] = self.api_client.select_header_accept(
-            ['application/json'])  # noqa: E501
-
-        # Authentication setting
-        auth_settings = ['basicAuth']  # noqa: E501
-
-        return self.api_client.call_api(
-            '/datasets/view/{ownerSlug}/{datasetSlug}', 'GET',
             path_params,
             query_params,
             header_params,