+++ kaggle/cli.py
@@ -445,11 +445,11 @@ def parse_datasets(subparsers):
         help=Help.command_datasets_files)
     parser_datasets_files_optional = parser_datasets_files._action_groups.pop()
     parser_datasets_files_optional.add_argument(
-        'dataset', help=Help.param_dataset)
+        'dataset', nargs='?', default=None, help=Help.param_dataset)
     parser_datasets_files_optional.add_argument(
         '-d',
         '--dataset',
-        dest='dataset',
+        dest='dataset_opt',
         required=False,
         help=argparse.SUPPRESS)
     parser_datasets_files_optional.add_argument(
@@ -469,7 +469,7 @@ def parse_datasets(subparsers):
     parser_datasets_download_optional = parser_datasets_download._action_groups.pop(
     )
     parser_datasets_download_optional.add_argument(
-        'dataset', help=Help.param_dataset)
+        'dataset', nargs='?', default=None, help=Help.param_dataset)
     parser_datasets_download_optional.add_argument(
         '-d',
         '--dataset',
@@ -629,11 +629,11 @@ def parse_datasets(subparsers):
     parser_datasets_metadata_optional = parser_datasets_metadata._action_groups.pop(
     )
     parser_datasets_metadata_optional.add_argument(
-        'dataset', help=Help.param_dataset)
+        'dataset', nargs='?', default=None, help=Help.param_dataset)
     parser_datasets_metadata_optional.add_argument(
         '-d',
         '--dataset',
-        dest='dataset',
+        dest='dataset_opt',
         required=False,
         help=argparse.SUPPRESS)
     parser_datasets_metadata_optional.add_argument(
@@ -655,11 +655,11 @@ def parse_datasets(subparsers):
     parser_datasets_status_optional = parser_datasets_status._action_groups.pop(
     )
     parser_datasets_status_optional.add_argument(
-        'dataset', help=Help.param_dataset)
+        'dataset', nargs='?', default=None, help=Help.param_dataset)
     parser_datasets_status_optional.add_argument(
         '-d',
         '--dataset',
-        dest='dataset',
+        dest='dataset_opt',
         required=False,
         help=argparse.SUPPRESS)
     parser_datasets_status._action_groups.append(
@@ -818,11 +818,11 @@ def parse_kernels(subparsers):
         help=Help.command_kernels_output)
     parser_kernels_output_optional = parser_kernels_output._action_groups.pop()
     parser_kernels_output_optional.add_argument(
-        'kernel', help=Help.param_kernel)
+        'kernel', nargs='?', default=None, help=Help.param_kernel)
     parser_kernels_output_optional.add_argument(
         '-k',
         '--kernel',
-        dest='kernel',
+        dest='kernel_opt',
         required=False,
         help=argparse.SUPPRESS)
     parser_kernels_output_optional.add_argument(
@@ -863,11 +863,11 @@ def parse_kernels(subparsers):
         help=Help.command_kernels_status)
     parser_kernels_status_optional = parser_kernels_status._action_groups.pop()
     parser_kernels_status_optional.add_argument(
-        'kernel', help=Help.param_kernel)
+        'kernel', nargs='?', default=None, help=Help.param_kernel)
     parser_kernels_status_optional.add_argument(
         '-k',
         '--kernel',
-        dest='kernel',
+        dest='kernel_opt',
         required=False,
         help=argparse.SUPPRESS)
     parser_kernels_status._action_groups.append(parser_kernels_status_optional)